resource-server: repositories
