+++ backend/resource_server/resource_server/src/main/kotlin/hu/bme/aut/resource_server/config/DbMigration.kt
@@ -10,6 +10,7 @@ class DbMigration {
     @Bean
     fun flywayMigration(): FlywayMigrationStrategy? {
         return FlywayMigrationStrategy { flyway: Flyway ->
+            flyway.baseline()
             flyway.repair()
             flyway.migrate()
         }
