+++ backend/resource_server/resource_server/src/main/kotlin/hu/bme/aut/resource_server/profile_snapshot/ProfileSnapshotItem.kt
@@ -1,5 +1,6 @@
-package hu.bme.aut.resource_server.user
+package hu.bme.aut.resource_server.profile_snapshot
 
+import hu.bme.aut.resource_server.ability.Ability
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
@@ -9,15 +10,15 @@ import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 
 @Entity
-data class ProfileItem(
+data class ProfileSnapshotItem (
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null,
 
     @ManyToOne
-    @JoinColumn(name="ability_id")
+    @JoinColumn(name = "ability_id")
     val ability: Ability,
 
     @Column
-    val abilityValue: Int
+    val abilityValue: Int,
 )
