+++ backend/resource_server/resource_server/src/main/kotlin/hu/bme/aut/resource_server/profile_snapshot/ProfileSnapshotRepository.kt
@@ -1,0 +1,10 @@
+package hu.bme.aut.resource_server.profile_snapshot
+
+import hu.bme.aut.resource_server.user.UserEntity
+import org.springframework.data.repository.CrudRepository
+import java.time.Instant
+
+interface ProfileSnapshotRepository : CrudRepository<ProfileSnapshot, Long> {
+    fun findAllByUser(user: UserEntity):List<ProfileSnapshot>
+    fun findAllByTimestampBetween(begin: Instant, end: Instant)
+}
