+++ backend/resource_server/resource_server/src/main/kotlin/hu/bme/aut/resource_server/user/UserRepository.kt
@@ -1,0 +1,8 @@
+package hu.bme.aut.resource_server.user
+
+import org.springframework.data.repository.CrudRepository
+import java.util.Optional
+
+interface UserRepository: CrudRepository<UserEntity, Int> {
+    fun findByUsername(username: String): Optional<UserEntity>
+}
