+++ backend/resource_server/resource_server/src/main/kotlin/hu/bme/aut/resource_server/user/role/RoleRepository.kt
@@ -1,0 +1,6 @@
+package hu.bme.aut.resource_server.user.role
+
+import org.springframework.data.repository.CrudRepository
+
+interface RoleRepository :CrudRepository<Role, Int>{
+}
