+++ backend/resource_server/resource_server/src/main/resources/application-test.yaml
@@ -1,0 +1,6 @@
+spring:
+  datasource:
+    url: jdbc:h2:mem:cogni-testDB;MODE=MYSQL;NON_KEYWORDS=USER
+    username: sa
+    password:
+    driver-class-name: org.h2.Driver
