+++ backend/resource_server/resource_server/src/main/resources/db/migration/V1_1__user_table.sql
