+++ backend/resource_server/resource_server/src/main/resources/db/migration/V1_2__game_table.sql
