+++ backend/resource_server/resource_server/src/main/resources/db/migration/V1_6__user_data_init.sql
